requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
certifi==2025.1.31
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
six==1.17.0
structlog==25.1.0
typing_extensions==4.12.2
tzdata==2025.1
yarl==1.18.3
requests==2.31.0
pyproj==3.6.1
shapely==2.0.2